+++ lib/ui/page/home/introduction/view.dart
@@ -18,13 +18,18 @@
 import 'package:animated_size_and_fade/animated_size_and_fade.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_animate/flutter_animate.dart';
 import 'package:get/get.dart';
 import 'package:share_plus/share_plus.dart';
 
+import '/domain/model/user.dart';
 import '/l10n/l10n.dart';
 import '/routes.dart';
 import '/themes.dart';
 import '/ui/page/home/widget/num.dart';
+import '/ui/page/login/controller.dart';
+import '/ui/page/login/view.dart';
+import '/ui/widget/download_button.dart';
 import '/ui/widget/modal_popup.dart';
 import '/ui/widget/outlined_rounded_button.dart';
 import '/ui/widget/svg/svg.dart';
@@ -51,8 +56,13 @@ class IntroductionView extends StatelessWidget {
     BuildContext context, {
     IntroductionViewStage initial = IntroductionViewStage.oneTime,
   }) {
+    final style = Theme.of(context).style;
+
     return ModalPopup.show(
       context: context,
+      background: initial == IntroductionViewStage.link
+          ? style.colors.background
+          : null,
       child: IntroductionView(initial: initial),
     );
   }
@@ -89,12 +99,13 @@ class IntroductionView extends StatelessWidget {
                     style: style.fonts.normal.regular.onPrimary,
                   ),
                 ),
+                const SizedBox(height: 16),
               ];
               break;
 
             case IntroductionViewStage.oneTime:
               header = ModalPopupHeader(
-                text: 'label_one_time_account_created'.l10n,
+                text: 'label_guest_account_created'.l10n,
               );
 
               children = [
@@ -137,7 +148,105 @@ class IntroductionView extends StatelessWidget {
                     style: style.fonts.normal.regular.onPrimary,
                   ),
                 ),
+                const SizedBox(height: 16),
               ];
+              break;
+
+            case IntroductionViewStage.link:
+              final guestButton = Center(
+                child: OutlinedRoundedButton(
+                  key: const Key('StartButton'),
+                  maxWidth: 290,
+                  height: 46,
+                  leading: Transform.translate(
+                    offset: const Offset(4, 0),
+                    child: const SvgIcon(SvgIcons.guest),
+                  ),
+                  onPressed: Navigator.of(context).pop,
+                  child: Text('btn_guest'.l10n),
+                ),
+              );
+
+              final signInButton = Center(
+                child: OutlinedRoundedButton(
+                  key: const Key('SignInButton'),
+                  maxWidth: 290,
+                  height: 46,
+                  leading: Transform.translate(
+                    offset: const Offset(4, 0),
+                    child: const SvgIcon(SvgIcons.enter),
+                  ),
+                  onPressed: () =>
+                      LoginView.show(context, initial: LoginViewStage.signIn),
+                  child: Text('btn_sign_in'.l10n),
+                ),
+              );
+
+              final applicationButton = Center(
+                child: OutlinedRoundedButton(
+                  key: const Key('DownloadButton'),
+                  maxWidth: 290,
+                  height: 46,
+                  leading: const Padding(
+                    padding: EdgeInsets.fromLTRB(4, 0, 0, 0),
+                    child: SvgIcon(SvgIcons.logo),
+                  ),
+                  onPressed: () => _download(context),
+                  child: Text('label_application'.l10n),
+                ),
+              );
+
+              if (PlatformUtils.isMobile) {
+                header = const ModalPopupHeader(dense: true);
+                children = [
+                  const SizedBox(height: 8),
+                  guestButton,
+                  const SizedBox(height: 15),
+                  signInButton,
+                  if (PlatformUtils.isWeb) ...[
+                    const SizedBox(height: 15),
+                    applicationButton,
+                  ],
+                  const SizedBox(height: 16),
+                  Center(
+                    child: Animate(
+                      effects: const [
+                        FadeEffect(
+                          delay: Duration(milliseconds: 0),
+                          duration: Duration(milliseconds: 2000),
+                        ),
+                        MoveEffect(
+                          delay: Duration(milliseconds: 0),
+                          begin: Offset(0, 100),
+                          end: Offset(0, 0),
+                          curve: Curves.ease,
+                          duration: Duration(milliseconds: 1000),
+                        ),
+                      ],
+                      child: Text(
+                        'label_messenger_by_gapopa'.l10n,
+                        style: style.fonts.normal.regular.secondary,
+                      ),
+                    ),
+                  ),
+                ];
+              } else {
+                header = ModalPopupHeader(
+                  text: 'label_messenger_by_gapopa'.l10n,
+                  dense: false,
+                );
+                children = [
+                  const SizedBox(height: 8),
+                  if (PlatformUtils.isWeb) ...[
+                    applicationButton,
+                    const SizedBox(height: 15),
+                  ],
+                  guestButton,
+                  const SizedBox(height: 15),
+                  signInButton,
+                  const SizedBox(height: 16),
+                ];
+              }
               break;
           }
 
@@ -153,9 +262,12 @@ class IntroductionView extends StatelessWidget {
                 physics: const ClampingScrollPhysics(),
                 children: [
                   header,
-                  ...children.map((e) =>
-                      Padding(padding: ModalPopup.padding(context), child: e)),
-                  const SizedBox(height: 16),
+                  ...children.map(
+                    (e) => Padding(
+                      padding: ModalPopup.padding(context),
+                      child: e,
+                    ),
+                  ),
                 ],
               ),
             ),
@@ -197,6 +309,40 @@ class IntroductionView extends StatelessWidget {
           ? const SvgIcon(SvgIcons.share)
           : const SvgIcon(SvgIcons.copy),
       label: 'label_your_direct_link'.l10n,
+    );
+  }
+
+  /// Opens a [ModalPopup] listing the buttons for downloading the application.
+  Future<void> _download(BuildContext context) async {
+    await ModalPopup.show(
+      context: context,
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          ModalPopupHeader(text: 'btn_download'.l10n),
+          const SizedBox(height: 12),
+          Flexible(
+            child: ListView(
+              padding: ModalPopup.padding(context),
+              shrinkWrap: true,
+              children: const [
+                DownloadButton.windows(),
+                SizedBox(height: 8),
+                DownloadButton.macos(),
+                SizedBox(height: 8),
+                DownloadButton.linux(),
+                SizedBox(height: 8),
+                DownloadButton.appStore(),
+                SizedBox(height: 8),
+                DownloadButton.googlePlay(),
+                SizedBox(height: 8),
+                DownloadButton.android(),
+              ],
+            ),
+          ),
+          const SizedBox(height: 18),
+        ],
+      ),
     );
   }
 }