+++ CHANGELOG.md
@@ -22,11 +22,25 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Media panel:
         - Redesigned incoming/outgoing call. ([#832], [#812])
+    - Chats tab:
+        - Direct link searching. ([#843], [#831])
+    - Home page:
+        - Redesigned introduction after following direct link. ([#848], [#820])
+
+### Fixed
+
+- Web:
+    - Inability to view camera, microphone and output devices in Firefox. ([#855])
 
 [#464]: /../../issues/464
 [#465]: /../../pull/465
 [#812]: /../../issues/812
+[#820]: /../../issues/820
+[#831]: /../../issues/831
 [#832]: /../../pull/832
+[#843]: /../../pull/843
+[#848]: /../../pull/848
+[#855]: /../../pull/855